backend/requirements.txt
# Web Framework
fastapi>=0.115.6
uvicorn[standard]>=0.34.0
python-multipart>=0.0.20

# Database
sqlalchemy>=2.0.36
aiosqlite>=0.20.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.1

# PII Detection
spacy>=3.8.7

# File Processing
pandas>=2.2.3
PyMuPDF>=1.25.3
python-docx>=1.1.2
sqlparse>=0.5.3

# Encryption
cryptography>=44.0.0

# Utilities
python-dotenv>=1.0.1
pydantic[email]>=2.10.4
